moved trajectory point to types.rs so planning -> control can use it
