+++ helios_core/src/types.rs
@@ -1,5 +1,6 @@
 // helios_core/src/types.rs
 
+use crate::frames::FrameAwareState;
 use nalgebra::{DVector, Isometry3};
 use serde::{Deserialize, Serialize};
 
@@ -42,4 +43,18 @@ pub trait TfProvider {
     /// Returns the world (ENU) pose of a frame directly from physics.
     /// Use this to bypass the estimator entirely (e.g. ground-truth mapping).
     fn world_pose(&self, frame: FrameHandle) -> Option<Isometry3<f64>>;
+}
+
+/// Reference trajectory point for the current time step.
+///
+/// Provided by the planner; consumed by path follower and controller.
+/// `state_dot` is `None` when the planner cannot or does not provide it
+/// (e.g. a waypoint planner). Feedforward controllers degrade gracefully
+/// to pure feedback and log a diagnostic when `state_dot` is absent.
+#[derive(Clone)]
+pub struct TrajectoryPoint {
+    pub state: FrameAwareState,
+    /// x_dot_ref = f(x_ref, u_nominal, t). None = feedforward unavailable.
+    pub state_dot: Option<DVector<f64>>,
+    pub time: f64,
 }
