+++ helios_sim/src/simulation/plugins/isolation/mock_path.rs
@@ -6,9 +6,9 @@
 // the path is already cached when the first control tick runs.
 
 use bevy::prelude::*;
-use helios_core::control::TrajectoryPoint;
 use helios_core::frames::{FrameAwareState, FrameId, StateVariable};
 use helios_core::planning::types::Path;
+use helios_core::types::TrajectoryPoint;
 use helios_runtime::stage::PipelineLevel;
 use serde::Deserialize;
